Update query to filter on statuses to prevent duplicates
Query filters on refund-pending and refund-requested status to prevent adding
multiple bulk_refund entries to the array for a payment. If the update modified
count is 0, it is logged instead of returning an error. This is to prevent the
operation from returning early and not processing legitimate refunds in the same
file.